// fiilter-angular/src/app/models/Fii.ts
export class Fii{
        codigo: string;
        setor: string;
        precoAtual: number;
        liquidez: number;
        dividendo: number;
        dy: number;
        dy12Media: number;
        pvpa: number;
        rentabilidadeTotal: number;
        vacanciaFisica: number;
        vacanciaFinanceira: number;
        qtdeAtivos: number;
}